+++ src/Game/GameObjects/BetonWall.cpp
@@ -68,6 +68,6 @@ void BetonWall::render() const
     renderBlock(EBlockLocation::BottomRight);
 }
 
-void BetonWall::update(const uint64_t delta)
+void BetonWall::update(const double delta)
 {
 }
