+++ src/Game/GameObjects/BetonWall.h
@@ -38,7 +38,7 @@ public:
 
     BetonWall(const EBetonWallType eBetonWallType, const glm::vec2& position, const glm::vec2& size, const float rotation, const float layer);
     virtual void render() const override;
-    virtual void update(const uint64_t delta) override;
+    virtual void update(const double delta) override;
 
 private:
     void renderBlock(const EBlockLocation eBlockLocation) const;
