+++ src/Game/GameObjects/Water.cpp
@@ -27,7 +27,7 @@ void Water::render() const
     renderBlock(EBlockLocation::BottomRight);
 }
 
-void Water::update(const uint64_t delta)
+void Water::update(const double delta)
 {
     m_spriteAnimator.update(delta);
 }
