+++ src/Renderer/Sprite.cpp
@@ -104,7 +104,7 @@ namespace RenderEngine {
         m_framesDescriptions = std::move(framesDescriptions);
     }
 
-    uint64_t Sprite::getFrameDuration(const size_t frameId) const
+    double Sprite::getFrameDuration(const size_t frameId) const
     {
         return m_framesDescriptions[frameId].duration;
     }
