+++ src/Renderer/SpriteAnimator.h
@@ -10,13 +10,13 @@ namespace RenderEngine {
     public:
         SpriteAnimator(std::shared_ptr<Sprite> pSprite);
         size_t getCurrentFrame() const { return m_currentFrame; }
-        void update(const uint64_t delta);
+        void update(const double delta);
 
     private:
         std::shared_ptr<Sprite> m_pSprite;
         size_t m_currentFrame;
-        uint64_t m_currentFrameDuration;
-        uint64_t m_currentAnimationTime;
+        double m_currentFrameDuration;
+        double m_currentAnimationTime;
     };
 
 }
